backend/project/requirements.txt
django
djangorestframework
markdown
django-filter
psycopg2-binary
django-mptt

drf-yasg
pymemcache
django-debug-toolbar

